# Build "optimize_for_inference" method